sample handler test case
